requirements.txt
setuptools==75.6.0
requests~=2.28.2
python-dotenv==1.0.1